requirements.txt
fastapi
uvicorn
pydantic
boto3
requests
python-dotenv
docling
celery
redis